+++ libs/model/src/index.ts
@@ -6,3 +6,4 @@ export * from './lib/organisation';
 export * from './lib/permissions';
 export * from './lib/contract';
 export * from './lib/event';
+export * from './lib/media';
